REFAC : remove question before launching IOTA2

+++ scripts/common/oso_directory.py
@@ -14,12 +14,26 @@
 #
 # =========================================================================
 
-import argparse,os,shutil,sys
 
-def GenerateDirectories(root):
+import os
+import shutil
+import serviceConfigFile as SCF
 
-    if os.path.exists(root) and root != "/":
+
+def GenerateDirectories(cfg):
+    """
+    generate IOTA2 output directories
+    """
+    if not isinstance(cfg, SCF.serviceConfigFile):
+        cfg = SCF.serviceConfigFile(cfg)
+
+    root = cfg.getParam('chain', 'outputPath')
+    rm_PathTEST = cfg.getParam("chain", "remove_outputPath")
+    start_step = cfg.getParam("chain", "startFromStep")
+
+    if os.path.exists(root) and root != "/" and rm_PathTEST and start_step == 1:
         shutil.rmtree(root,ignore_errors=False)
+
     os.mkdir(root)
     if os.path.exists(root+"/model"):
         shutil.rmtree(root+"/model")
@@ -62,17 +76,6 @@ def GenerateDirectories(root):
     os.mkdir(root+"/cmd/features")
     os.mkdir(root+"/cmd/fusion")
     os.mkdir(root+"/cmd/splitShape")
-
-if __name__ == "__main__":
-
-    parser = argparse.ArgumentParser(description = "This function creates directories for classifications")
-    parser.add_argument("-root",dest = "root",help ="path where all directories will be create",required=True)
-    args = parser.parse_args()
-
-    GenerateDirectories(args.root)
-
-
-
